Extract org.wildfly.clustering.infinispan.client module from org.wildfly.clustering.infinispan.spi
