+++ clustering/infinispan/client/pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
   ~ JBoss, Home of Professional Open Source.
-  ~ Copyright 2010, Red Hat, Inc., and individual contributors
+  ~ Copyright 2018, Red Hat, Inc., and individual contributors
   ~ as indicated by the @author tags. See the copyright.txt file in the
   ~ distribution for a full listing of individual contributors.
   ~
@@ -28,7 +28,7 @@
 
     <parent>
         <groupId>org.wildfly</groupId>
-        <artifactId>wildfly-clustering</artifactId>
+        <artifactId>wildfly-clustering-infinispan</artifactId>
         <!--
         Maintain separation between the artifact id and the version to help prevent
         merge conflicts between commits changing the GA and those changing the V.
@@ -36,14 +36,19 @@
         <version>17.0.0.Beta1-SNAPSHOT</version>
     </parent>
 
-    <artifactId>wildfly-clustering-infinispan</artifactId>
-    <packaging>pom</packaging>
+    <artifactId>wildfly-clustering-infinispan-client</artifactId>
+    <packaging>jar</packaging>
 
-    <name>WildFly: Infinispan modules</name>
+    <name>WildFly: Infinispan Client SPI</name>
 
-    <modules>
-        <module>client</module>
-        <module>spi</module>
-        <module>extension</module>
-    </modules>
+    <dependencies>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>wildfly-clustering-common</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.infinispan</groupId>
+            <artifactId>infinispan-client-hotrod</artifactId>
+        </dependency>
+    </dependencies>
 </project>
