+++ clustering/infinispan/extension/src/main/java/org/jboss/as/clustering/infinispan/subsystem/remote/RemoteCacheContainerResourceDefinition.java
@@ -48,7 +48,7 @@ import org.jboss.as.controller.transform.description.AttributeConverter.DefaultV
 import org.jboss.as.controller.transform.description.ResourceTransformationDescriptionBuilder;
 import org.jboss.dmr.ModelNode;
 import org.jboss.dmr.ModelType;
-import org.wildfly.clustering.infinispan.spi.InfinispanRequirement;
+import org.wildfly.clustering.infinispan.client.InfinispanClientRequirement;
 import org.wildfly.clustering.service.UnaryRequirement;
 
 /**
@@ -65,8 +65,8 @@ public class RemoteCacheContainerResourceDefinition extends ChildResourceDefinit
     }
 
     public enum Capability implements CapabilityProvider {
-        CONTAINER(InfinispanRequirement.REMOTE_CONTAINER),
-        CONFIGURATION(InfinispanRequirement.REMOTE_CONTAINER_CONFIGURATION),
+        CONTAINER(InfinispanClientRequirement.REMOTE_CONTAINER),
+        CONFIGURATION(InfinispanClientRequirement.REMOTE_CONTAINER_CONFIGURATION),
         ;
 
         private final org.jboss.as.clustering.controller.Capability capability;
