+++ feature-pack/src/main/resources/modules/system/layers/base/org/wildfly/clustering/infinispan/client/main/module.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
+
 <!--
   ~ JBoss, Home of Professional Open Source.
   ~ Copyright 2010, Red Hat, Inc., and individual contributors
@@ -21,29 +22,22 @@
   ~ 02110-1301 USA, or see the FSF site: http://www.fsf.org.
   -->
 
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
+<module xmlns="urn:jboss:module:1.5" name="org.wildfly.clustering.infinispan.client">
+    <properties>
+        <property name="jboss.api" value="private"/>
+    </properties>
 
-    <parent>
-        <groupId>org.wildfly</groupId>
-        <artifactId>wildfly-clustering</artifactId>
-        <!--
-        Maintain separation between the artifact id and the version to help prevent
-        merge conflicts between commits changing the GA and those changing the V.
-        -->
-        <version>17.0.0.Beta1-SNAPSHOT</version>
-    </parent>
+    <resources>
+        <artifact name="${org.wildfly:wildfly-clustering-infinispan-client}"/>
+    </resources>
 
-    <artifactId>wildfly-clustering-infinispan</artifactId>
-    <packaging>pom</packaging>
-
-    <name>WildFly: Infinispan modules</name>
-
-    <modules>
-        <module>client</module>
-        <module>spi</module>
-        <module>extension</module>
-    </modules>
-</project>
+    <dependencies>
+        <module name="javax.api"/>
+        <module name="org.infinispan.client.hotrod"/>
+        <module name="org.infinispan.commons"/>
+        <module name="org.jboss.as.clustering.common"/>
+        <module name="org.jboss.as.controller"/>
+        <module name="org.wildfly.clustering.service"/>
+        <module name="org.jboss.msc"/>
+    </dependencies>
+</module>
