+++ feature-pack/src/main/resources/modules/system/layers/base/org/wildfly/extension/clustering/web/main/module.xml
@@ -47,6 +47,7 @@
         <module name="org.jboss.msc"/>
         <module name="org.jboss.staxmapper"/>
         <module name="org.jboss.vfs"/>
+        <module name="org.wildfly.clustering.infinispan.client"/>
         <module name="org.wildfly.clustering.infinispan.spi"/>
         <module name="org.wildfly.clustering.service"/>
         <module name="org.wildfly.clustering.web.infinispan"/>
